+++ src/test/java/integration/TestObject.java
@@ -1,0 +1,17 @@
+package integration;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Dummy Test object used for the integration tests
+ * 
+ * @author bmary
+ * 
+ */
+public class TestObject {
+
+    public final String name = "test object";
+    public final int points = 12;
+    public List<String> properties = Arrays.asList("prop1", "prop2");
+}